Split construction_anim.rs into construction_anim/ subdirectory
Split crates/rendering/src/construction_anim.rs (636 lines) into focused modules
under construction_anim/ for better maintainability:

- types.rs (34 lines): Components and resource types
- meshes.rs (270 lines): Mesh builders and asset initialization
- systems.rs (221 lines): All ECS systems
- tests.rs (120 lines): Unit tests
- mod.rs (13 lines): Thin facade with re-exports

All public items re-exported from mod.rs — no external import changes needed.
